apps/data-management-api/packages/shared-config/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shared-config"
version = "0.1.0"
description = "Shared configuration utilities for the Vecinita platform"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["shared_config"]

[tool.ruff]
line-length = 100
